+++ src/lib.rs
@@ -15,7 +15,7 @@
 //!
 //! // Parse existing config
 //! let path = r"/path/to/df/data/init/init.txt";
-//! let mut conf = Config::read_str(read_to_string(path).unwrap());
+//! let mut conf = Config::read_str(read_to_string(path)?);
 //!
 //! // Read some value
 //! let sound = conf.get("SOUND");
@@ -28,6 +28,8 @@
 
 #[macro_use]
 extern crate lazy_static;
+
+use std::collections::HashMap;
 
 use regex::Regex;
 
@@ -196,6 +198,17 @@ impl Config {
         })
     }
 
+    /// Returns an iterator over (`key`, `value`) tuples.
+    pub fn keys_values_iter(&self) -> impl Iterator<Item = (&str, &str)> + '_ {
+        self.lines.iter().filter_map(|x| {
+            if let Line::Entry(entry) = x {
+                Some((entry.get_key(), entry.get_value()))
+            } else {
+                None
+            }
+        })
+    }
+
     /// Returns the string representing the configuration in its current state (aka what you'd write to the file usually).
     pub fn print(&self) -> String {
         let mut buff = Vec::<String>::with_capacity(self.lines.len());
@@ -210,6 +223,16 @@ impl Config {
         }
 
         buff.join("\r\n")
+    }
+}
+
+impl From<Config> for HashMap<String, String> {
+    fn from(conf: Config) -> Self {
+        let mut output = HashMap::new();
+        conf.keys_values_iter().for_each(|(key, value)| {
+            output.insert(key.to_owned(), value.to_owned());
+        });
+        output
     }
 }
 
@@ -351,5 +374,31 @@ mod tests {
         assert_eq!(conf.len(), 2);
         assert_eq!(conf.get(&a), None);
         assert_eq!(conf.remove(random_alphanumeric()), 0);
+    }
+
+    #[test]
+    fn test_keys_values_iter() {
+        let a: String = random_alphanumeric();
+        let b: String = random_alphanumeric();
+        let mut conf = Config::new();
+        conf.set(&a, "foo");
+        conf.set(&b, "bar");
+        let mut iter = conf.keys_values_iter();
+        assert_eq!(Some((a.as_ref(), "foo")), iter.next());
+        assert_eq!(Some((b.as_ref(), "bar")), iter.next());
+        assert_eq!(None, iter.next());
+    }
+
+    #[test]
+    fn test_hashmap() {
+        let a: String = random_alphanumeric();
+        let b: String = random_alphanumeric();
+        let mut conf = Config::new();
+        conf.set(&a, "foo");
+        conf.set(&b, "bar");
+        let hash_map_owned: HashMap<String, String> = conf.into();
+        assert_eq!(hash_map_owned.len(), 2);
+        assert_eq!(hash_map_owned.get(&a).unwrap(), "foo");
+        assert_eq!(hash_map_owned.get(&b).unwrap(), "bar");
     }
 }