+++ changelog.md
@@ -1,3 +1,6 @@
-dfconfig v0.1.0
-===============
+# dfconfig future
+* Added `Config::test_keys_values_iter()`
+* Implemented `From<Config> for HashMap<String, String>`
+
+# dfconfig v0.1.0
 * First version, basic functionality to read/write DF `init.txt`-like files.